+++ pyccel/stdlib/STC_Extensions/Dict_extensions.h
@@ -30,5 +30,19 @@ static inline i_val _c_MEMB(_pop_with_default)(Self* self, i_keyraw rkey, i_val 
     return _c_MEMB(_pop)(self, rkey);
 }
 
+/**
+ * Remove and returns a (key, value) pair from the dictionary.
+ *
+ * @param[inout] self : A pointer to the dictionary instance.
+ * @param[out] key : A key.
+ * @param[out] val : The associated value.
+ */
+static inline void _c_MEMB(_popitem)(Self* self, i_key* key, i_val* val) {
+    _m_iter iter = _c_MEMB(_begin)(self);
+    *key = iter.ref->first;
+    *val = iter.ref->second;
+    _c_MEMB(_erase_at)(self, iter);
+}
+
 #include <stc/priv/template2.h>
 #undef _i_is_map