+++ tests/epyccel/modules/generic_functions.py
@@ -34,7 +34,7 @@ def tmplt_head_1(x : 'Z', y : 'Z'):
     return x + y
 
 @template('O', types=['int', 'real'])
-def local_overide_1(x : 'O', y : 'O'):
+def local_override_1(x : 'O', y : 'O'):
     return x + y
 
 @template('Z', types=['int', 'real'])
@@ -87,9 +87,9 @@ def tst_tmplt_head_1():
     y = tmplt_head_1(5.5, 7.3)
     return x, y
 
-def tst_local_overide_1():
-    x = local_overide_1(5, 4)
-    y = local_overide_1(6.56, 3.3)
+def tst_local_override_1():
+    x = local_override_1(5, 4)
+    y = local_override_1(6.56, 3.3)
     return x, y
 
 def tst_tmplt_tmplt_1():